create and use col_ct, row_ct and dim_ct

+++ benchmark/benchmark/binary_op/size/Eigen_ct_stack.cpp
@@ -26,14 +26,14 @@ void bm(benchmark::State& state, Op op, D1, D2){
 
 	using namespace Eigen;
 
-	Matrix< T, D1::ct_rows, D1::ct_cols > m1;
+	Matrix< T, size_t(D1::ct_rows), size_t(D1::ct_cols) > m1;
 	for(int y = 0; y < m1.rows(); ++y){
 		for(int x = 0; x < m1.cols(); ++x){
 			m1(y, x) = dis(gen);
 		}
 	}
 
-	Matrix< T, D2::ct_rows, D2::ct_cols > m2;
+	Matrix< T, size_t(D2::ct_rows), size_t(D2::ct_cols) > m2;
 	for(int y = 0; y < m2.rows(); ++y){
 		for(int x = 0; x < m2.cols(); ++x){
 			m2(y, x) = dis(gen);
